requirements.txt
Django==5.1.3
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-extensions
python-dotenv
gunicorn
whitenoise
psycopg2-binary==2.9.10 # Use psycopg for PostgreSQL (or replace with `mysqlclient` for MySQL)
asgiref==3.8.1
sqlparse==0.5.2


uvicorn
fastapi  # Optional: If using FastAPI for microservices
